+++ benches/basic.rs
@@ -17,6 +17,8 @@ pub fn bench(c: &mut Criterion) {
     )] = &[
         ("hash/rapidhash", Box::new(vector::bench_rapidhash), Box::new(int::bench_rapidhash), Box::new(object::bench_rapidhash)),
         ("hash/rapidhash_raw", Box::new(vector::bench_rapidhash_raw), Box::new(int::bench_rapidhash_raw), Box::new(object::bench_rapidhash)),
+        ("hash/rapidhash_cc_v1", Box::new(vector::bench_rapidhash_cc_v1), Box::new(int::bench_rapidhash_cc_v1), Box::new(object::bench_rapidhash)),
+        ("hash/rapidhash_cc_v2", Box::new(vector::bench_rapidhash_cc_v2), Box::new(int::bench_rapidhash_cc_v2), Box::new(object::bench_rapidhash)),
         ("hash/default", Box::new(vector::bench_default), Box::new(int::bench_default), Box::new(object::bench_default)),
         ("hash/fxhash", Box::new(vector::bench_fxhash), Box::new(int::bench_fxhash), Box::new(object::bench_fxhash)),
         ("hash/gxhash", Box::new(vector::bench_gxhash), Box::new(int::bench_gxhash), Box::new(object::bench_gxhash)),
@@ -30,9 +32,10 @@ pub fn bench(c: &mut Criterion) {
         ("hash/farmhash", Box::new(vector::bench_farmhash), Box::new(int::bench_farmhash), Box::new(object::bench_farmhash)),
         ("hash/highwayhash", Box::new(vector::bench_highwayhash), Box::new(int::bench_highwayhash), Box::new(object::bench_highwayhash)),
         ("hash/rustc-hash", Box::new(vector::bench_rustchash), Box::new(int::bench_rustchash), Box::new(object::bench_rustchash)),
+        ("hash/foldhash", Box::new(vector::bench_foldhash), Box::new(int::bench_foldhash), Box::new(object::bench_foldhash)),
     ];
 
-    let sizes = [2usize, 8, 16, 64, 100, 177, 256, 1024, 4096];
+    let sizes = [2usize, 8, 16, 25, 50, 64, 160, 256, 1024, 4096];
 
     for (name, string_fn, int_fn, object_fn) in groups.into_iter() {
         let mut group = c.benchmark_group(name.to_string());