+++ benches/int.rs
@@ -1,4 +1,4 @@
-use std::hash::Hasher;
+use std::hash::{BuildHasher, Hasher};
 use criterion::Bencher;
 use rapidhash::RAPID_SEED;
 
@@ -68,6 +68,26 @@ pub fn bench_rapidhash_raw() -> Box<dyn FnMut(&mut Bencher)> {
             rand::random::<u64>()
         }, |i: u64| {
             rapidhash::rapidhash_inline(&i.to_le_bytes(), RAPID_SEED)
+        }, criterion::BatchSize::SmallInput);
+    })
+}
+
+pub fn bench_rapidhash_cc_v1() -> Box<dyn FnMut(&mut Bencher)> {
+    Box::new(move |b: &mut Bencher| {
+        b.iter_batched(|| {
+            rand::random::<u64>()
+        }, |i: u64| {
+            rapidhash_c::rapidhashcc_v1(&i.to_le_bytes(), 0xbdd89aa982704029)
+        }, criterion::BatchSize::SmallInput);
+    })
+}
+
+pub fn bench_rapidhash_cc_v2() -> Box<dyn FnMut(&mut Bencher)> {
+    Box::new(move |b: &mut Bencher| {
+        b.iter_batched(|| {
+            rand::random::<u64>()
+        }, |i: u64| {
+            rapidhash_c::rapidhashcc_v2(&i.to_le_bytes(), RAPID_SEED)
         }, criterion::BatchSize::SmallInput);
     })
 }
@@ -220,6 +240,18 @@ pub fn bench_rustchash() -> Box<dyn FnMut(&mut Bencher)> {
             rand::random::<u64>()
         }, |i: u64| {
             let mut hasher = rustc_hash::FxHasher::default();
+            hasher.write_u64(i);
+            hasher.finish()
+        }, criterion::BatchSize::SmallInput);
+    })
+}
+
+pub fn bench_foldhash() -> Box<dyn FnMut(&mut Bencher)> {
+    Box::new(move |b: &mut Bencher| {
+        b.iter_batched(|| {
+            rand::random::<u64>()
+        }, |i: u64| {
+            let mut hasher = foldhash::quality::RandomState::default().build_hasher();
             hasher.write_u64(i);
             hasher.finish()
         }, criterion::BatchSize::SmallInput);