+++ rapidhash-c/cpp/rapidhash_v1_wrapper.cpp
@@ -1,0 +1,1 @@
+#include "rapidhash_v1.h"