+++ rapidhash-c/cpp/rapidhash_v2_wrapper.cpp
@@ -1,0 +1,1 @@
+#include "rapidhash_v2.h"