+++ rapidhash-c/src/lib.rs
@@ -1,0 +1,5 @@
+mod rapidhash_v1_bindings;
+mod rapidhash_v2_bindings;
+
+pub use rapidhash_v1_bindings::*;
+pub use rapidhash_v2_bindings::*;